#5389: deleted others.py and moved the operations it contained into the right places
